ticket [https://trac.osgeo.org/mapguide/ticket/1421] OGC WFS 1.1.0 Support
11. Test wfs:wfs-1.1.0-Basic-GetFeature-tc23.1 (s0001/d1e35137_1/d1e740_1/d1e25217_1/d1e6182_1)

Assertion: The following query parameters are required for all KVP-encoded service requests except !GetCapabilities: service, request, version.

For !GetFeature requests, if service or version parameter is not specified, MapGuide should response a WFS exception.

12. Test wfs:wfs-1.1.0-Basic-GetFeature-tc44.2 (s0001/d1e35137_1/d1e740_1/d1e25217_1/d1e6238_1)

Assertion: Version number specified in a given request must correspond to a version supported by the service.

For !GetFeature requests, if the version number specified in the request is not supported, MapGuide should generate a WFS exception.
  
13. Test wfs:wfs-1.1.0-Basic-GetFeature-tc32.1 (s0001/d1e35137_1/d1e740_1/d1e25217_1/d1e6226_1)

Assertion: An ogc:Literal element may include simple or complex content. If a child element appears it may belong to any target namespace, but a top-level declaration for the item must be available.

For property names in filter, there will be a prefix like "gml:" or user specified prefix. Those prefixes should be removed to make sure properties can be found by FDO API.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5191 a442ad82-b40c-0410-ae44-e6d7bd6e4a36