On behalf of: Spark Liu
Fix problem that totally opaque watermark will not be rendered.

Reason: To optimize watermark rendering, there is code to jump over if watermark is totally transparent. However, the logic about judging "totally transparent" is actually "totally opaque" in the incorrect code.

Fix way: change the logic to the correct way.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5237 a442ad82-b40c-0410-ae44-e6d7bd6e4a36