Ticket [https://trac.osgeo.org/mapguide/ticket/1458] Incorrect initial view of published WMS layers
This defect is caused by the parent layer in the GetCapabilities response document. Because the parent layer's boundingbox is hard coded to (-90,-180,90,180), so while the children layers inherit from the parent layer, their default boundingbox will be set to the parent layer's boudingbox. 

This submission also adds CITEWmsEnabled to handle the OGC WMS CITE test requests.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5266 a442ad82-b40c-0410-ae44-e6d7bd6e4a36