MapGuide Trac# 1456:
While the GeometryApiGen.xml file was correctly updated, the RFC 94 submission 
did not correctly update the MapGuideApiGen.xml file thus the generation of the 
Java version of the API failed.  Thee associated ..Constant.xml files were 
similarly handled.

Neither were the Makefile.am files for Java and Php updated to include the new 
interfaces.  Correcting that issue lead to the discovery that the  
implementation of RFC 76  (Grids and MGRS) did not include several new classes 
in the Makefile.am files.  This situation was corrected as well.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@5270 a442ad82-b40c-0410-ae44-e6d7bd6e4a36