Ticket [https://trac.osgeo.org/mapguide/ticket/1462] New CRSs cannot be supported in WMS service
All published layer headers use SRS to indicate the coordinate system which is a standard of WMS 1.1.0. But WMS 1.3.0 uses CRS to define the coordinate system. 

Because there's no CRS defined in the published layer header, parent layer's CRS will be used as all children layers' CRS, which is EPSG:4326. 

I update all the CRS to SRS in the template file Server/src/Wms/1.3.0.xml.awd for the backward compatibility issue. 

This submission also fixes a test failure for OGC CITE WMS 1.1.0. I update the refer of the DTD which is required by OGC. 

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5278 a442ad82-b40c-0410-ae44-e6d7bd6e4a36