Ticket [https://trac.osgeo.org/mapguide/ticket/1236] "Sticky" tooltips
Submit on behalf of jng

Currently tooltips will always follow the mouse pointer, as a result if there is any interactive content inside the tooltip, it is impossible to actually interact with the said content using the mouse. 

What the tooltip should ideally do is to stay at the position it was first displayed, and only disappear when the mouse pointer leaves the tooltip's bounding box. This is how it's done in Fusion and allows for clicking of any interactive content inside the tooltip. 



git-svn-id: http://svn.osgeo.org/mapguide/trunk@5305 a442ad82-b40c-0410-ae44-e6d7bd6e4a36