On Behalf of: Spark Liu
Fixes for several watermark rendering problems:
1. Watermark will be opaque when it's set totally transparent.
Fix the incorrect "return" code when not rendering at all.

2. Watermark will be rendered as selected feature when some features selected.
This is because originally rendering selected features is the last step of rendering. So it doesn't set selection mode to false after it has done selected features rendering.
This submission changes this.

3. Opacity will be doubled in the viewer when some features are selected.
This is because watermark is also rendered in the image of selected features. Change the code to only render watermark when layer is rendered. As in viewer layer is never rendered along with selected feature, this problem will be gone.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5365 a442ad82-b40c-0410-ae44-e6d7bd6e4a36