Ticket [https://trac.osgeo.org/mapguide/ticket/1520] Bounding Box value doesn't affect the view of WMS Server
If a request contains a BBox whose area does not overlap at all with the <BoundingBox> element in the service metadata for the requested layer, the server shall return empty content (that is, a blank map or an graphic elemenet file with no elements) for that map. Any features that are partly or entirely contained in the Bounding Box shall be returned in the appropriate format. 

According to OGC WMS specification, MapGuide should apply a spatial filter on the request layers. Features out of the <Boundingbox> which specified by users should not appear on the WMS layer. 

This might bring MapGuide some performace issues, so I also optimized the process: if the request boundingbox within the user specified boundingbox, no spatial filter will be applied. 

In order to apply spatial filter on wms layers, I created wms layers in session repository and updated their layer definitions. 




git-svn-id: http://svn.osgeo.org/mapguide/trunk@5400 a442ad82-b40c-0410-ae44-e6d7bd6e4a36