Follow up submission for trac ticket 1538.
When reading all available definitions from the dictionaries, all code names have been takes "as-is". Unfortunately, the CsMap's coordsys.dsc contains some coordinate systems that reference datums via lowercased strings - but the corresponding datums in datums.dsc contain uppercased versions of these strings. For example, "Afgooye.LL" references "Afgooye" as its datum but that has the key "AFGOOYE" in the dictionary. Changed the code in MentorDictionary.h so that the std::maps are filled with lower cased strings only and also the lookup into these std::maps is done with lowercased versions of the key strings (changes in CoordSys.cpp and CoordSysDatum.cpp).

Reviewed by Hugues.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5432 a442ad82-b40c-0410-ae44-e6d7bd6e4a36