Fixed two bugs which were causing crashes when round-tripping data with To/FromAgf().
One was that MultiLineStrings were serialized incorrectly (number of geometries was not correctly set).

The second was that m_dimensionality was blindly set to what comes in with an AGF stream, which sometimes may contain M values. During serialization, this dimensionality value was written back to the output stream, but no M values were written since LineBuffer does not support those. This fix was to remove the M flag from m_dimensionality.

Also changed some FdoDimensionality types to int, to reduce the API dependency on FDO types, but this is just a small part of a larger job.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@5436 a442ad82-b40c-0410-ae44-e6d7bd6e4a36