Fix for ticket 1577 (http://trac.osgeo.org/mapguide/ticket/1577)
The CCoordinateSystemMathComparator::SameCoordinateSystem incorrectly returned false for 2 systems, if both weren't referencing a datum or an ellipsoid.
This is a valid definition though - coordinate systems with projection type non-earth (NERTH) don't have either set.

Modified the [SameCoordinateSystem] method so it does also check the case where both systems neither reference a datum nor an ellipsoid. If the projection type is then NERTH, the method will do the remaining checks, e.g. check the x/y offsets on equality + the unit.

Reviewed by Hugues.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@5477 a442ad82-b40c-0410-ae44-e6d7bd6e4a36