Fix for ticket 1576 "Values set via MgCoordinateSystem::SetLonLatBounds() are not written back to CsMap's dictionary for user defined systems" (http://trac.osgeo.org/mapguide/ticket/1576)
The MgCoordinateSystem class has an API "SetLonLatBounds". This allows for dynamically setting a range in which coordinates are valid for the respective system. If switched on (by default, or by calling this method with a set of valid parameters), CsMap will fail or at least set a warning, when a coordinate transformation is done with input values that do not fall within the range specified.

The API did only set the values passed in on the internally held [cs_Csprm_] struct (that's being used for conversion) - and not on the coordinate system definition, that's being saved to disk. Since CsMap 13.00, most of the default dictionary entries have valid ranges set. Before this submission, it was simply not possible to store valid ranges with a custom coordinate system definition.

There's only 1 code change in [CCoordinateSystem::SetLonLatBounds()]: The min/max lat/long values are now also set on the [m_csprm.csdef.ll_*] struct so that this information also gets written to disk when the (custom) coordinate system definition is stored. I've also added a documentation for the methods in [MgCoordinateSystem] that deal with the min/max lat/long stuff.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5478 a442ad82-b40c-0410-ae44-e6d7bd6e4a36