Follow up submission for ticket #1577 (http://trac.osgeo.org/mapguide/ticket/1577)
While the original fix itself was correct it was missing the Nrthsrt ("Nerth with scale and rotation") projection type.

Modified the code in CoordSysMathComparator.cpp so it does now also checks for [MgCoordinateSystemProjectionCode::Nrthsrt] in case a coordinate system does neither reference a datum nor an ellipsoid.

At the moment, the check will still return false, if the 2 systems being compared are of projection type [Nerth] and [Nrthsrt] but have equal scale and rotation parameters set. This is due to the fact, that the code does also compare the projection code itself on equality: This is already commented in the code before this submission. So, I won't change that now. This requires some more effort that's not related to ticket 1577.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@5489 a442ad82-b40c-0410-ae44-e6d7bd6e4a36