This submission simply corrects an ASSERT condition in LineBuffer.cpp.  The LoadFromAgf method has:
    for (int q=0; q<num_geoms; ++q)
    {
        // skip past geometry type of subgeometry
        // we know it is LineString or Polygon or Point respectively
        if (is_multi) *ireader++;

        // read cordinate typeB
        FdoDimensionality dim = (FdoDimensionality)*ireader++;

        // ensure that all dimensionalities of each geometry are the same
        _ASSERT(q==0 || m_dimensionality == dim);

        m_dimensionality = dim & ~FdoDimensionality_M; //LineBuffer doesn't support M


In the case where you have a multi-linestring of XYM data, the first pass (q==0)
will skip the assert and m_dimensionality is set to FdoDimensionality_XY (M bit
is supressed).  In the second pass (q==1), dim will be FdoDimensionality_XYM,
but m_dimensionality will be FdoDimensionality_XY and the assert will fire.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@5520 a442ad82-b40c-0410-ae44-e6d7bd6e4a36