Fix Ticket 1607
This is an edge case that if a composite style contains only labels, it is not rendered correctly. All the labels are draws in the bottom-left corner.

This happens because when laying out the labels using eight-surrounding algorithm, it calls GetLastSymbolExtent to get the corresponding symbol extent as the base coordinate. However, in the case that this composite type style contains only labels, this LastSymbolExtent was never set. Thus, it'll use (0,0) as base coordinate of the algorithm while it should use the centroid of the geometry.

So I made this change that it'll check the returned value of GetLastSymbolExtent. if this value is zero, we'll use the centroid of the geometry instead.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5569 a442ad82-b40c-0410-ae44-e6d7bd6e4a36