[https://trac.osgeo.org/mapguide/ticket/1598]: Support editing of point styles based on W2D symbols in a Symbol Library. Unlike the original description, this was never actually implemented. So now we've done that. However previews of such point styles are not perfect. It will only show the symbol and does not consider any color overrides that the style has applied. Nevertheless, seeing a symbol (albeit maybe of the wrong color) is better than seeing nothing.
Also modify the generated LayerDefintion xml classes to fix an incorrectly generated W2D symbol reference class

Also modify the layer preview to use either the name of the feature source (new layer) or the name of the edited resource itself (existing layer) as the legend label for the preview map. This is much more readable than an ugly guid.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@5726 a442ad82-b40c-0410-ae44-e6d7bd6e4a36