On behalf of Sam Wang. This submission is the first part of implementing RFC 114: Extend MapGuide API to Support Save Points.
Three new interface methods have been added to MgTransaction interface as 
proposed in the RFC. All the concrete class that inherits this interface 
currently provides corresponding implementation.

In order that the new API can be called from both MG server API and MG web API, 
we add three corresponding methods in MgFeatureService interface. These three 
methods are almost the same with that of MgTransaction, except that they pass in 
a transaction id to idetify which transaction they will operate on.  

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@5966 a442ad82-b40c-0410-ae44-e6d7bd6e4a36