This submission i for Ticket [https://trac.osgeo.org/mapguide/ticket/1754]
This is to implement RFC 113 ( http://trac.osgeo.org/mapguide/wiki/MapGuideRfc113) .

The new schema of SymbolDefinition?, layer definition map definition and watermark definition is added. the new version numbers are all 2.4.0. Also, the Stylizer is modified to handle the additional scale properties. In evaluate method of SE_Polygon and SE_Polyline, the line buffer is scaled for each feature.

The MdfParser? and MdfModel? is also update to version 2.4.0 for the above 4 schemas The MdfModel? server unit test was updated to test the new schema versions

TESTING PERFORMED: Tested in MapGuide using mapagent to set a layer definition with scale factors being a constant & expression. Both work. In the case of expression, I use featureid as the width, witch varies per-feature.

The MdfModel? unittest was run and passed.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@5996 a442ad82-b40c-0410-ae44-e6d7bd6e4a36