On behalf of Spark Liu: Fix for ticket 1756: ServerFeatureService.DescribeSchemaAsXml will throw exception when geometry property is inherited from base class
When the geometry property is inherited from base class (which means no extra geometry property defined in child class), ServerFeatureService.DescribeSchemaAsXml will throw exception.
This is because when trying to change MgClassDefinition back to FdoClassDefinition, it always try to get geometry property from the property collection in child class instead of base class.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@5999 a442ad82-b40c-0410-ae44-e6d7bd6e4a36