Ticket [https://trac.osgeo.org/mapguide/ticket/1737] (Submit on behalf of Sam Wang)
This submission implements MapGuide RFC 119: Enhance URL information in vector layer definition.
To change the layer definition, corresponding Mdf model and parser code has also to be changed so that the new version of layer definition(mdf model) and parser can be compatible with the old version without problem.
Unit tests have be run to verify that.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@6043 a442ad82-b40c-0410-ae44-e6d7bd6e4a36