Submit on behalf of Spark Liu: Fix for ticket 1782 - Customized function such as "EXTENT" can't be run on SQL Server Spatial
When coming to customized function such as "EXTENT", MapGuide server will try to calculate by itself instead of using FDO provider. So it will request the actual property instead of the expression into FDO. When getting properties from reader, it first check if the return column number is only 1, and use the first column as return column.

However, there is no guarantee in FDO interface that the return reader only contains the requesting column. For example, SQL Server spatial adds PK column as well, and many code may have been written depending on that.

So to make the code robust, if there are more than one columns returned, we should parse the expression when we gets the reader and only use the column we need.

The fix only happens when the original codes fail, so it will not affect original logic.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@6075 a442ad82-b40c-0410-ae44-e6d7bd6e4a36