RFC 116 Update - Remove extra interface
There were several revisions made to the original RFC 116 prior to its submission for approval.  Due to these revisions, a previously proposed, but then retracted, interface (IsReentrant) was removed.  This interface, which is not specified in the approved RFC, was erroneously included in the RFC 116 submission.  This submission removes this function from the interface to the MgCoordinateSystemTransform object.  

Note that the GctTestA.cpp and GctTestSupport.cpp modules are a part of the GeometryConsoleTest application which is only used internally to test changes to the MgGeometry.dll independently of any of the several vertical applications which use it.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@6093 a442ad82-b40c-0410-ae44-e6d7bd6e4a36