Fix stylization memory leak
SE_RenderStyle objects were getting leaked in the case where you selected features
which had labels based on enhanced stylization.

Stylizing a label involves the following:
  * clone the SE_RenderStyle used for the label
  * put the cloned style into an SE_LabelInfo object
  * call SE_Renderer:ProcessSELabelGroup with the SE_LabelInfo
During regular rendering the label renderer takes ownership of the cloned style and
eventually deletes it.  But during selection rendering the label renderer is not
invoked, and therefore the cloned style is leaked.

To fix this I added a destructor to SE_LabelInfo which deletes any referenced style.
In the selection rendering case the destructor will now automatically delete the
style.  In the regular rendering case the label renderer still takes ownership of
the cloned style, and so we simply clear the 'style' member variable in SE_LabelInfo
to prevent the destructor from deleting it.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@6149 a442ad82-b40c-0410-ae44-e6d7bd6e4a36