This submission is mainly to address ticket 1824 (http://trac.osgeo.org/mapguide/ticket/1824).
In RFC116, API has been added to the MgCoordinateSystemTransform class allowing to retrieve the actual geodetic transformation steps that are performed when converting from the source coordinate system.

I'm applying some corrections to that last change:
--> Documentation:
The description of the [NumberOfGeodeticTransformations] left the reader unclear in what exactly this method returns in which cases and what exactly is considered a NULL-transformation
at MapGuide API level. For example, setting up an explicit NULL-transformation from one datum to another results in [NumberOfGeodeticTransformations] returning a non-zero count.

Also fixed the [GetGeodeticTransformation]'s comment so it reflects what the method is doing. In particular, I corrected the statement, that a zero-index value would always return a NULL pointer what's not the case.

--> Implementation: 
The [GetGeodeticTransformation] accessed an empty std::wstring via the index operator - this resulted in an assertion in debug builds. The behavior in release would be undefined in the best case.
Changed the implementation so it checks the index parameter before it tries to do anything with it.

--> API naming:
All new API added for RFC116 is named 
[...GeodeticTransformation...()]. Just the [NumberOfGeodeticTransformations] was named differently. Changed that to [GetGeodeticTransformationCount] so that we now a consistent naming scheme.

[GetGeodeticTransformationCount()]
[GetGeodeticTransformation(INT32)]
[GetGeodeticTransformationDirection(INT32)]

Note though, that I'm not changing the GeometryConsoleTest project here as this has the code that would have exercised parts of the [...GeodeticTransformation...] API excluded from building.

Changes reviewed by Hugues.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@6166 a442ad82-b40c-0410-ae44-e6d7bd6e4a36