This submission fixes ticket 1825 (http://trac.osgeo.org/mapguide/ticket/1825).
The new API
[SetCompareInternalDatumOldParameters] and [GetCompareInternalDatumOldParameters] has been added to the [MgCoordinateSystemMathComparator] class.

These methods allow an API client to disable the comparison of geodetic transformation information when comparing 2 datums. Due to the changes done in RFC94 and the CSMAP RFC2, comparing the geodetic transformation directly on the datum is no longer necessary as the information is retrieved from a separate dictionary by CSMAP.

With this submission, the default behavior is kept so that a new math comparator object would still check these properties by default.

See also the addendum section for RFC94.

Changes reviewed by Hugues.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@6167 a442ad82-b40c-0410-ae44-e6d7bd6e4a36