Fix ticket 1828 (http://trac.osgeo.org/mapguide/ticket/1828).
Calling CCoordinateSystemGeodeticTransformDef::SerializeFrom() did result in a heap corruption in case the de-serialized CSMAP struct was found to be invalid. In that case, the code free'd the [this->transformDefinition] pointer - what again was being tried to do in the Destructor later on. This resulted in a heap corruption.

Fixed the code so it does only free the temporarily allocated memory if an exception was thrown.
In case of success, the method does only free the previously [this->transformDefinition] pointer and keep the newly allocated block.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@6173 a442ad82-b40c-0410-ae44-e6d7bd6e4a36