Fix http://trac.osgeo.org/mapguide/ticket/1831.
The ::IsValid() checks performed when deserializing an MgCoordinateSystemGeodeticTransformationDef or MgCoordinateSystemGeodeticPath object from a CSMAP struct were too restrictive. Both did always require the datums to exist in the current dictionaries.

With that inconsistent behavior it was impossible to create a new datum and an accompanying transformation or path on the fly, because the code would always insist on the datum to be available. For exactly that reason, all other MgCoordinateSystem definition types like cs, datums and ellipsoids never required any "child object" to exist when (de-)serializing. 

In both classes, I've removed the cs_*CHK_* flags that were passed to the respective CSMAP chk() methods. CSMAP's "check" methods are now called with simply 0 instead - in which case CSMAP performs some basic validation only but doesn't require the datums to exist already in the dictionaries.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@6175 a442ad82-b40c-0410-ae44-e6d7bd6e4a36