[https://trac.osgeo.org/mapguide/ticket/1854]: Implement RFC123. This submission takes advantage of recently introduced FDO Join APIs to provide a high-performance optimization shortcut for Feature Joins if:
 a) The extended feature class joins against another feature class from the same feature source
 b) The feature source supports the FDO Join APIs. As of FDO 3.7, SQLite and SQL Server satisfy this condition.

Although use of the FDO expression engine is generally avoided where possible due to performance, we have no choice but to use the FdoExpressionEngineUtilDataReader here in order to support SpatialExtents(). The reason we need to support SpatialExtents() on extended feature classes is in order for the Feature Source Preview to generate a usable MBR for previewing data.

Also included are unit tests to benchmark and test some SQLite-based feature sources set up to take advantage of this optimization shortcut. The 2 SQLite data stores used (and included in this submission) are:

 1. The Sheboygan Parcels data set, split into 2 parts:
   - The autogenerated id, geometry and join key (ParcelFeatures)
   - The remaining attributes, including the join key (Parcels)
 2. A hand-built data set of US/Australian countries/states/cities

See the RFC123 link for full details.

Building in VS2010 will be momentarily broken (as I do not have access to VS2010 on hand to update its vcxproj files)

git-svn-id: http://svn.osgeo.org/mapguide/trunk@6239 a442ad82-b40c-0410-ae44-e6d7bd6e4a36