Fix ticket [https://trac.osgeo.org/mapguide/ticket/1865] - showTaskPane is called multiple times when switching maps
showTaskPane will be called multiple times for loading a map after a map has been switching for several times.

This is because index page register MAP_LOAD event to register TASK_PANE_LOADED event. After MAP_LOAD event is triggered several times, the same function is registered to TASK_PANE_LOADED event with exactly same number. This is not as expected. We just want showTaskPane be called once when TASK_PANE_LOADED is triggered for one time.

We should remove the MAP_LOAD event registering after TASK_PANE_LOADED is registered.

It seems registering TASK_PANE_LOADED directly when fusion is initialized is a better way. But there is a slice difference between fusion loaded and map loaded. If registering in fusion loaded, task pane will always show when map loaded. But if registering in map loaded, task pane will show AFTER map is loaded for the first time. Maybe the original submission wants the latter way. So we don't change this thing.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@6246 a442ad82-b40c-0410-ae44-e6d7bd6e4a36