This submission fixes ticket 1862 "CCoordinateSystemEnumInteger32 leaks memory when being destructed" (http://trac.osgeo.org/mapguide/ticket/1862)
The main problem is, that the CCoordinateSystemEnumInteger32 class did not delete its [m_pnValues] array which holds all the Integer32 values upon destruction.

Added the delete[] call to the destructor.
Also added a call to the [ClearFilter] so that any such (auto-ptr'ed) values are also released.

Changes reviewed by Walt.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@6283 a442ad82-b40c-0410-ae44-e6d7bd6e4a36