on behalf of libr (Bruce li)
Reason:
There are many handlers for the Event - MAP_LOADED of Map, one of them deregister itself after executed while other handlers have not executed, this action disturb the sequence of the execution of handlers behind.

solution:
Remove the code of deregistering Eventhandler. take another way to disable the Eventhandler.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@6434 a442ad82-b40c-0410-ae44-e6d7bd6e4a36