On behalf of Andy Zhang. Fix ticket #1960 in trunk: ByteSource and ByteReader memory leak in Java Api
The generated Java method MgByteSource::GetReader() is like below:
public MgByteReader GetReader() throws MgException  {
    long cPtr = MapGuideJavaApiJNI.MgByteSource_GetReader(swigCPtr);
    return (cPtr == 0) ? null : (MgByteReader)ObjectFactory.createObject(MapGuideJavaApiJNI.getClassId(cPtr), cPtr, false);
  }

The last parameter indicated whether the object owns memory itself. In Java API, it is always false. That means it will not call MapGuideJavaApiJNI.delete_MgByteReader(swigCPtr) when deleting the object, which results in memory leak.

Now change the last parameter to 'true' to fix this issue. The change is consistant with CSharp.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@6611 a442ad82-b40c-0410-ae44-e6d7bd6e4a36