[https://trac.osgeo.org/mapguide/ticket/2073]: Add a second semi-workaround for mg-desktop's tiled map limitation. We introduce the following new viewer properties:
 * UseRenderMapIfTiledLayersExist
 * RespectFiniteDisplayScales

If UseRenderMapIfTiledLayersExist is true and the map has tiled layers, we use the RenderMap API instead of RenderDynamicOverlay, which has the benefit of being able to render tiled layers if they are visible. If RespectFiniteDisplayScales is true, the viewer will snap any zoom to the map's finite scale list (like AJAX/Fusion does). The two properties combined, will approximate the AJAX/Fusion viewer behaviour wrt a Tiled Map.

The reason this is another semi-workaround and not the definitive solution is that this approach still does not use (our already implemented) tile service APIs and cannot take advantage of pre-rendered tiles in the tile cache, but is more approachable than the existing workaround (using ConvertTiledGroupsToNonTiled) in that the official MapGuide behaviour for RenderMap is the same. Anywhere where we can have behavioural consistency between MapGuide and mg-desktop is a good thing.

Long story short. We can now view tiled maps with minimal compromise, but nothing is actually tiled and cached (the *actual* reason to use a tiled map). Everything is still dynamic. Ideally we want to have the viewer leverage and render off of the existing tile service APIs, and then be able to tweak this logic to support external tiled maps (eg. Google/Bing/OSM). This is still a pipe-dream, but the primary objective of being able to at least *view* MapGuide Tiled Maps has been achieved.

The AppLayout engine has also been updated to support these new viewer properties and also to support a map: lookup property for Component Definitions. Basically any component property that starts with a map: prefix will use the MapDefinition of the named map, thus avoiding having to hard-code a Map Definition. The <Map> element now includes a mandatory <Name> element the is the name of the map. This name is validated on AppLayout load to ensure no bad characters are included.

git-svn-id: http://svn.osgeo.org/mapguide/branches/2.4@7064 a442ad82-b40c-0410-ae44-e6d7bd6e4a36