Support MentorUserDictionaryPath in serverconfig.ini (part of RFC 127)
Implemented support for the [MentorUserDictionaryPath] setting in the serverconfig.ini file.
If that setting is available (i.e. not an empty string), the server startup code will call into
the MgCoordinateSystemCatalog's SetUserDictionaryDir() method thus overriding
the default location of the user dictionary path - which may or may not have been set before.

While an invalid [MentorDictionaryPath] setting leads to the server to failing to startup, an invalid
[MentorUserDictionaryPath] does not. It will be logged as a warning though.

Changes reviewed by Jackie. 

git-svn-id: http://svn.osgeo.org/mapguide/trunk@7111 a442ad82-b40c-0410-ae44-e6d7bd6e4a36