Have MgCoordinateSystemCatalog::AreDictionaryFilesWritable() to check [m_sUserDir], too (part of RFC 127)
The [MgCoordinateSystemCatalog::AreDictionaryFilesWritable()] method checks, whether updating definitions in any of CS-Map's dictionary can be done.

Due to the implementation of RFC127, we've to make sure, that [AreDictionaryFilesWritable] checks the alternative user directory for writing access, i.e. 
[m_sUserDir] instead of [m_sDir]. But only, if [SetUserDictionaryDir()] had been successfully called before. Because all updates will end up in modifying the files in
[m_sUserDir] anyway.

I've now changed the implementation of [AreDictionaryFilesWritable] to check whether [m_sUserDir] can be written to (if set).
If not available, the code behaves as before, i.e. it will check, whether all CSD files can be updated in [m_sDir].

Added the documentation to the CoordinateSystemCatalog.h header.

Changes reviewed by Hugues.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@7200 a442ad82-b40c-0410-ae44-e6d7bd6e4a36