Fix #2192 (MgMultiGeometry::Deserialize doesn't work with MgAgfStream)
I've updated the code in MgGeometry::Deserialize which peeks for the geometry type to account for
the type of the passed-in MgStream.   Only if it's a pure MgStream do we use the code added in r3502.
Otherwise we use the original pre-3502 code.

I also updated the server's geometry unit test to exercise serialization / deserialization for all
geometry types in the context of an MgAgfReaderWriter.  This validates the fix.  All geometry unit
tests are passing.


git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@7238 a442ad82-b40c-0410-ae44-e6d7bd6e4a36