[Submitted on behalf of Steven Xu.  Reviewed by myself.]
Fix ticket [https://trac.osgeo.org/mapguide/ticket/2197]: The usage of "swscanf" results in bad performance in MgStylization.dll.
http://trac.osgeo.org/mapguide/ticket/2197

The function "swscanf" is used in the method "SE_StyleVisitor::ParseDoublePair?" and "SE_StyleVisitor::ParseGeometry?", and it's invoked many times in cases where there are many symbols and layers, and it results in bad performance.

This submission fixes this problem.  Now we use a quick algorithm to parse the geometry string in these two methods.  If this quick algorithm fails, parse it again with "swscanf".


git-svn-id: http://svn.osgeo.org/mapguide/trunk@7240 a442ad82-b40c-0410-ae44-e6d7bd6e4a36