Submit on behalf of Andy Zhang Integrate 7253 to trunk.
Fix ticket [https://trac.osgeo.org/mapguide/ticket/2195] http://trac.osgeo.org/mapguide/ticket/2195

The behavior of system call _wstat in VC9 and VC10 is different when handling file which size is bigger than 2^31. In VC9, it returns 0. But in VC10, it returns -1. Then some checks will be failed when opening a big file. 
To fix this issue, we need to use _wstat64 instead of _wstat. 



git-svn-id: http://svn.osgeo.org/mapguide/trunk@7258 a442ad82-b40c-0410-ae44-e6d7bd6e4a36