Submit on behalf of Andy Zhang. Fix ticket [https://trac.osgeo.org/mapguide/ticket/2331]: http://trac.osgeo.org/mapguide/ticket/2331 : Select polygon hangs when digitizing an invalid polygon.
This error happens on some data sources such as PostGis? 2.0. If only digitize 2 vertice when select polygon, all the following operations don't have response any more. We have to restart to Server to make it work again. 
To fix the issue, we need to check the vertex count before sending the query.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@7712 a442ad82-b40c-0410-ae44-e6d7bd6e4a36