Merge in streaming work from streaming_v2 sandbox. This includes the following changes:
MgReader streaming refactoring

 - dllexport the ByteSourceImpl class from the Foundation library
 - Implement a new MgReaderByteSourceImpl class that inherits from ByteSourceImpl. It takes a MgReader and mime type as input and provides an adapter that allows a MgReader to be streamed out as a MgByteReader
 - Modify MgHttpReaderStreamer to work off of a MgByteReader instead of a MgReader
 - For MgByteReader responses that should be chunked out, the implement operation writes the Transfer-Encoding = Chunked setting to the MgHttpHeader of the response
 - The CGI/ISAPI/Apache handlers have been modified to
    - No longer handle MgReader responses as they are now wrapped by a MgByteReader
    - If handling a MgByteReader, check for the Transfer-Encoding = Chunked setting in the MgHttpHeader of the response. If the setting exists, forward the byte reader to the modified Apache/Isapi/CgiReaderStreamer that now knows how to chunk output a MgByteReader. Otherwise, they will output the byte reader as normal.

[https://trac.osgeo.org/mapguide/ticket/1070]: Stream WFS GetFeature responses from the mapagent. We do this by bypassing the OGC XML templating code.
 
MgFeatureService.GetWfsFeature() is already returning WFS GetFeature response content. I do not know why this response then had to go through the OGC XML templating code, where the full XML content had to be read into memory (the cause of the memory spikes on large GetFeature responses). The only difference between the GetFeature response (that goes through the XML templating code) and the GetFeature response (that bypasses the XML templating code) is that the GetFeature response that goes through the XML templating code contains comment markers denoting each individual feature in the response. Besides that, the responses are identical!

For large responses there will be a small pause because the server-side is still buffered first before being streamed down to the web tier. The problem here is that MgServerFeatureService.GetWfsFeature() is dumping the GML result into a temporary file before returning a MgByteReader from this temp file. Ideally for the server-side to stream this as well, we should have a custom FdoIoStream class that writes direct to the TCP/IP MgStream with some type gymnastics so that we can still return a MgByteReader to respect the original API.

This submission renders the MgWfsFeatures class and some methods of MgOgcWfsServer useless. This submission removes these items in question. Also GetFeature-related template fragments in the .awd files are no longer used and so they have been removed as well.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@7782 a442ad82-b40c-0410-ae44-e6d7bd6e4a36