Implement MapGuide RFC 141
Merged revision(s) 8256-8259 from sandbox/jng/geos34x:
Remove old GEOS and replace it with GEOS 3.4.2. As it seems to be convention here, we're checking in as "geos" and not "geos-3.4.2" so that future Oem component updates won't need to involve lots of path fixing.
........
Update projects and code for GEOS 3.4.2
 - Fix geometry exception macros due to semantic changes in exception handling. The API of GEOSException has changed and it now inherits from std::exception and is caught by const reference
 - Order sensitivity of sub-geometries in the resulting geometry of a SymetricDifference operation. The MgGeometry::SymetricDifference API has been documented with this change of behaviour.
 - Fix a decimal precision issue for a area-related test case. That test case was updated to use CPPUNIT_ASSERT_DOUBLES_EQUAL to compare up to a certain number of decimal places.
........
Update build_oem.sh for geos. Also add silent rules support for geos
........
Add remark about the change of behaviour in SymetricDifference due to the geos upgrade.
........


git-svn-id: http://svn.osgeo.org/mapguide/trunk@8345 a442ad82-b40c-0410-ae44-e6d7bd6e4a36